requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.15.0
python-pptx>=0.6.21
openpyxl>=3.1.2